+++ sdk/identity/Azure.Identity.Broker/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Other Changes
 
-- Updated `Azure.Identity` to 1.14.2 to apply a security fix in the updated Microsoft.Identity.Client` dependency.
+- Updated `Azure.Identity` to 1.14.2 to apply a security fix in the updated `Microsoft.Identity.Client` dependency.
 
 ## 1.3.0-beta.3 (2025-06-10)