+++ sdk/identity/Azure.Identity.Broker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 1.2.1
+
+### Other Changes
+
+- Updated `Microsoft.Identity.Client.Broker` dependency to version 4.73.1
+
 ## 1.3.0-beta.4 (2025-07-11)
 
 ### Features Added